+++ CHANGELOG.md
@@ -1,12 +1,22 @@
 # Changelog
 
+## [0.0.1-exp.5] - 2025-11-03
+
+Implemented initial version of thread and group folding.
+
+### Fixed so it's eaiser to click on small events.
+
 ## [0.0.1-exp.4] - 2025-08-04
 
 ### Fixed so it's eaiser to click on small events.
 
 ## [0.0.1-exp.3] - 2025-07-30
 
-### Fixed so event focus works on in-active frames.
+* Fixed so event focus works on in-active frames.
+* Updated Burst dependency to version 1.8.11
+* Inspector UI
+* Updated the `com.unity.burst` dependency to version `1.8.24`
+* Updated the Burst dependency to version 1.8.25
 
 ## [0.0.1-exp.2] - 2025-07-29